Add files via upload
My second data science project took me on a fascinating journey - this time with the company Eniac. As they grappled with a key business question about whether it was wise to offer more discounts, I turned to data storytelling using Python and Pandas to provide the answer.

Through this experience, I honed my skills at cleaning and transforming raw, unruly data, demonstrating that valuable insights can be found even in the midst of chaos. Armed with Python's Matplotlib & Seaborn libraries, I learned to weave the cleaned data into compelling visual narratives. These visuals served as the backbone of a business proposal that not only communicated insights, but also inspired actionable decisions.

This project underscored that data isn't just a collection of numbers; it's a living story waiting to be told. By answering Eniac's critical question, I demonstrated that the art of data storytelling can be a powerful tool for driving business success.